* Source/NSUserDefaults.m: * Source/NSConnection.m: * Source/NSCharacterSet.m: * Tools/make_strings/StringsFile.m: * Tools/pl.m: Fix instances where NSNotFound is stored in or compared to an unsigned or int variable.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@34884 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSConnection.m
@@ -1583,7 +1583,7 @@ static NSLock	*cached_proxies_gate = nil;
   GS_M_LOCK(IrefGate);
   if (IrunLoops != nil)
     {
-      unsigned	pos = [IrunLoops indexOfObjectIdenticalTo: loop];
+      NSUInteger	pos = [IrunLoops indexOfObjectIdenticalTo: loop];
 
       if (pos != NSNotFound)
 	{
